+++ libs/internal/include/launchdarkly/data_model/flag.hpp
@@ -83,5 +83,12 @@ struct Flag {
     bool trackEvents;
     bool trackEventsFallthrough;
     std::optional<std::uint64_t> debugEventsUntilDate;
+
+    /**
+     * Returns the flag's version. Satisfies ItemDescriptor template
+     * constraints.
+     * @return Version of this flag.
+     */
+    [[nodiscard]] inline std::uint64_t Version() const { return version; }
 };
 }  // namespace launchdarkly::data_model