+++ libs/internal/tests/data_model_serialization_test.cpp
@@ -13,7 +13,8 @@ TEST(SDKDataSetTests, DeserializesEmptyDataSet) {
         boost::json::value_to<tl::expected<data_model::SDKDataSet, JsonError>>(
             boost::json::parse("{}"));
     ASSERT_TRUE(result);
-    ASSERT_FALSE(result->segments);
+    ASSERT_TRUE(result->segments.empty());
+    ASSERT_TRUE(result->flags.empty());
 }
 
 TEST(SDKDataSetTests, ErrorOnInvalidSchema) {
@@ -29,8 +30,15 @@ TEST(SDKDataSetTests, DeserializesZeroSegments) {
         boost::json::value_to<tl::expected<data_model::SDKDataSet, JsonError>>(
             boost::json::parse(R"({"segments":{}})"));
     ASSERT_TRUE(result);
-    ASSERT_TRUE(result->segments);
-    ASSERT_TRUE(result->segments->empty());
+    ASSERT_TRUE(result->segments.empty());
+}
+
+TEST(SDKDataSetTests, DeserializesZeroFlags) {
+    auto result =
+        boost::json::value_to<tl::expected<data_model::SDKDataSet, JsonError>>(
+            boost::json::parse(R"({"flags":{}})"));
+    ASSERT_TRUE(result);
+    ASSERT_TRUE(result->flags.empty());
 }
 
 TEST(SegmentTests, DeserializesMinimumValid) {